Refactor: Move Repository test -> Service Test
 - Move Repository test -> Service Test for division layer work
